requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
black==24.2.0
celery==5.3.6
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.43.0
colorama==0.4.6
coverage==7.4.1
cron-descriptor==1.4.5
dj-database-url==2.1.0
Django==4.2.10
django-appconf==1.1.0
django-axes==6.2.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cleanup==8.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.12.0
django-filter==23.5
django-imagekit==5.0.0
django-ratelimit==4.1.0
django-storages==1.14.2
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.0.0
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
kombu==5.5.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pilkit==3.0
pillow==10.2.0
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pytest==7.4.4
pytest-django==4.7.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
